Rename P2P unit test files and folders
